requirements.txt
Flask==3.0.0
python-frontmatter==1.1.0
PyYAML==6.0.1
Markdown==3.5.1
